+++ html5valdialect/src/test/java/com/github/masa_kunikata/html5val/util/MockURLBuilder.java
@@ -1,4 +1,4 @@
-package com.msd_sk.thymeleaf.dialects.html5val.util;
+package com.github.masa_kunikata.html5val.util;
 
 public class MockURLBuilder {
